make a utilities folder and move bcrypt into it
