+++ src/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { Users } from 'src/entities/user.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BcryptService } from 'src/bcrypt/bcrypt.utility';
+import { BcryptService } from 'src/utilities/bcrypt/bcrypt.utility';
 import { AuthModule } from 'src/auth/auth.module';
 
 @Module({
