+++ src/user/user.service.ts
@@ -5,7 +5,7 @@ import {
   InternalServerErrorException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { BcryptService } from 'src/bcrypt/bcrypt.utility';
+import { BcryptService } from 'src/utilities/bcrypt/bcrypt.utility';
 import { Users } from 'src/entities/user.entity';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
